requirements.txt
anyio==3.6.2
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.94.0
Flask==2.2.3
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
motor==3.1.1
mypy-extensions==1.0.0
numpy==1.24.2
orjson==3.8.7
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
platformdirs==3.1.1
plotly==5.13.1
psutil==5.9.4
pydantic==1.10.6
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.0.post1
tenacity==8.2.2
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.3
